Trabalho_Prático/requirements.txt
Flask==3.0.0
python-dotenv==0.18.0
